requirements.txt
Adafruit-Blinka==8.37.0
adafruit-circuitpython-74hc595==1.4.4
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-charlcd==3.4.10
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-ds18x20==1.4.0
adafruit-circuitpython-mcp230xx==2.5.13
adafruit-circuitpython-onewire==2.0.7
adafruit-circuitpython-requests==3.2.2
adafruit-circuitpython-typing==1.10.3
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
Adafruit-Python-DHT==1.4.0
black==24.4.1
click==8.1.7
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pika==1.3.2
platformdirs==4.2.1
psutil==5.9.8
pyftdi==0.55.0
pypng==0.20220715.0
pyserial==3.5
python-dotenv==1.0.1
pyusb==1.2.1
PyYAML==6.0.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
spidev==3.6
sysv-ipc==1.1.0
tomli==2.0.1
typing-extensions==4.10.0
